requirements.txt
Django==4.2.5
Pillow==10.0.0
djangorestframework==3.14.0
django-extensions==3.2.3